Consolidate reporting and UI update
