+++ .gitignore
@@ -27,8 +27,6 @@ venv/
 
 # Sample/mockup files (if any)
 samples/
-RR_IMPACT_ANALYSIS.md
-DESIGN.md
 
 #Drawio files
 *.drawio
@@ -37,3 +35,6 @@ DESIGN.md
 *.temp
 *.bkp
 *.bk
+
+# Test output
+**/test_reports/
